Uninstall support added
We use a brute force approach to find what files an installed added to
the environment, then store those along with the package environment in
the local package database. This lets us remove all the files a package
placed.  Some corner cases not handled here, but enough to get going.