Tests for jump and XPM_ROOT variable
The jump command can now take arbitrary commands to run instead of just
running bash.  This is used in the test to check that variables are being
properly set.  Also a marker variable XPM_ROOT is being set. This will be
used by other commands to remove the need for the explicit passing of that
command.