Fix for install path re-writing
If we had one binary string with the install dir in it twice, the re-write
logic would fail, because it didn't take into account the offset shift
caused by previous operations.

Here have re-written the detection such that each string gets one list,
with all the specified offsets, and then the null terminator location. The
replacement logic uses this to properly handle this case.