Use pre-load to force environment ld-linux usage (HACK)
Each ELF binary has the path to its dynamic loader hard coded into the
INTERP section of the binary.  On 64bit linux it's something like:

  /lib64/ld-linux-x86-64.so.2

If we are using are own version of libc, it will come with it's own
version of that file, which will conflict.  So we use LD_PRELOAD to force
our version of that library to be loaded instead.  It works for now
but will probably need todo some kind path-rewriting in the future.