Make XPKG_TREE and XPKG_REPO to be paths
This allows for the user to have multiple repositories or trees of packages
at once.  It uses a very simple algorithm of looking through the repos in
order until the package version combination if found.  If a general package
query is done, the newest version in any repo is returned.