Partial fix for uClibc dynamic linking issue
This fixes not finding the ll_tzname symbol, but now it doesn't even find
stdout, which is just crazy.  Attempts to debug the by enabled
SUPPORT_LD_DEBUG did not show anything helpful.  It just showed a proper
list of symbols (like stdout, malloc, free, etc.) and nothing being found.
It did show the finding of the uClibc, and it does contain the needed
symbols.  The dynamic linker just can't find them.