Basic ld.so indirection for Linux
In this setup we make sure there is always a lib/ld-linux-xpkg.so in our
environment, and that binaries built have their interp pointed their.
This is currently done with a build flag but if needed we can patch the
binaries.

This indirection link points to local ld.so in the environment if their is
one otherwise it points to the system ld.so. This allows for ld changes and
custom ld's without any other changes to the packages.