Update Toolsets to contain environment variable definitions
We need to set things like the CC and CXX environment variable differently
depending on the toolset.  So that is now done as part of the toolset
definition.  This means the ld.so indirection is done through adjusting the
LDFLAGS environment variable.

This fixes the toolset test to actually compile the desired package with
the toolset in questions.  It also makes the toolset include libc (it was
not previously).