Perl 5.20.0 practical extraction and report language
The old unix standby of a scripting language.  Used by several projects
during their build process.  The patch we need is because perl records the
length of the embedded INC and will error out if it finds null padding at
the end.  Since we insert padding when move around the perl install this
breaks thing, our patch just truncates the length to match.