Change install database format to store file data separately and lazy load
Loading all the files for every package each time a command was run is
quite slow.  This new setup caches the file data in JSON files and only
loads them when needed.  This provides a decent speedup vs. the old system
but we still need to update the repo layer to only load file data as needed
too.